feat: Add Selective GAPIC support for TypeScript generation (#1690)
This PR adds support for Selective GAPIC generation through the service yaml. 

For example, the service yaml below would now be supported via changes in this PR: 
```
publishing:
  typescript_settings:
    common:
      selective_gapic_generation:
        methods:
        - google.cloud.aiplatform.v1.PredictionService.Predict
        - google.cloud.aiplatform.v1.JobService.CreateBatchPredictionJob
        - google.cloud.aiplatform.v1.JobService.CreateCustomJob
       generate_omitted_as_internal: true
	 as_deny_list: true
      destinations:
       - PACKAGE_MANAGER
  python_settings:
    common:
      destinations:
       - PACKAGE_MANAGER
```

In this PR, we make updates to the service yaml ingestion as well as the way we augment methods in order to categorize the methods by an enum. `HIDDEN` methods are not generated at all in the generator. For `INTERNAL` methods, the templates can then generate @internal annotations in the JSDocs and add appropriate comments for methods that should not be used on the final client layer.